add 个人中心界面修改 add 关于我们添加 add 问题反馈界面添加 add 添加分享功能

+++ module-search/src/main/java/com/rm/module_search/bean/SearchResultBean.kt
@@ -10,11 +10,11 @@ import com.rm.business_lib.db.download.DownloadAudio
  *
  */
 data class SearchResultBean(
-    var audio: Long,//搜索到的书籍总量
+    var audio: Int,//搜索到的书籍总量
     var audio_list: List<DownloadAudio>,//书籍列表
-    var member: Long,//搜索到的主播数量
+    var member: Int,//搜索到的主播数量
     var member_list: List<MemberBean>,//主播列表
-    var sheet: Long,//搜索到的听单数量
+    var sheet: Int,//搜索到的听单数量
     var sheet_list: List<SearchSheetBean>//听单列表
 )
 
@@ -22,15 +22,15 @@ data class MemberBean(
     var avatar: String,//头像
     var avatar_url: String,//头像url
     var fans: String,//粉丝数
-    var is_follow: Long,//是否已经关注，0:未关注；1:已经关注，未登陆默认返回0
-    var level: Long,//等级
+    var is_follow: Int,//是否已经关注，0:未关注；1:已经关注，未登陆默认返回0
+    var level: Int,//等级
     var member_id: String,//主播用户id
     var nickname: String,//主播名
     var member_intro: String//简介
 )
 
 data class SearchSheetBean(
-    var audio_num: Long,//书籍数量
+    var audio_num: Int,//书籍数量
     var created_at: String,//创建时间
     var favour_num: Int,//收藏人数
     var sheet_cover: String,//封面
